backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "bcrypt>=4.0.1,<5.0.0",
    "email-validator>=2.0.0",
]
